Websocket work for URL
